Improve behavior on environment without openssl library
Check OpenSSL::VERSION constant to see if openssl library is available
or not.

Due to how autoload works, "defined?(OpenSSL)" does not always work
as expected: it reports as "constant" even if openssl library actually
fails to load. net-smtp eagerly loads openssl, but another library can
register autoload and it will still be a problem.

Test cases are adjusted to skip relevant tests when openssl library is
not available.