remove SMTPUTF8RequiredError
Email addresses that contain non-ASCII characters no longer cause an error if the server does not support SMTPUTF8.
This is because even if the server does not support SMTPUTF8, it may not cause an error for email addresses with non-ASCII characters. The server should decide whether to cause an error or not.